@@ -4,13 +4,16 @@
   "private": true,
   "homepage": "http://ritobritade.github.io/",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.0.1"
+    "@react-three/drei": "9.53.0",
+    "@react-three/fiber": "8.10.0",
+    "@react-three/postprocessing": "2.7.0",
+    "@types/three": "0.148.0",
+    "leva": "0.9.34",
+    "r3f-perf": "6.6.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-scripts": "5.0.1",
+    "three": "0.148.0"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -39,6 +42,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.3"
+    "gh-pages": "^3.2.3",
+    "prettier": "2.3.2"
   }
 }
